Paths, paths, paths...

+++ tools/devops/publish-performance-data.sh
@@ -2,27 +2,14 @@
 
 DATA_REPO_NAME=xamarin-macios-data
 
-# grab Azure Devop's authorization token from the current repo, and add it to the global git configuration
-#AUTH=$(git config -l | grep AUTHORIZATION | head -1 | sed 's/.*AUTHORIZATION: //')
-#AUTH_MD5=$(echo "$AUTH" | md5)
-#git config --global http.extraheader "AUTHORIZATION: $AUTH"
-#echo "AUTH_MD5=$AUTH_MD5"
-
-# Debug spew, checking if the authorization token is correct
-#git ls-remote https://github.com/xamarin/maccore || true
-#git ls-remote https://github.com/xamarin/xamarin-macios-data || true
-
-#pushd .
-#cd ../maccore
-#git fetch
-#popd
-
-
-#git clone https://github.com/xamarin/$DATA_REPO_NAME
-
 DIR=$DATA_REPO_NAME/perf-data/samples/$BUILD_SOURCEBRANCHNAME/$BUILD_SOURCEVERSION/$SYSTEM_JOBID
 mkdir -p "$DIR"
 
+XMLS=(xamarin-macios/tests/sampletester/bin/Debug/tmp-test-dir/execution-logs/*.xml)
+if ! test -f "${XMLS[0]}"; then
+	echo "##vso[task.logissue type=warning]Could not find any performance data to publish"
+	exit 0
+fi
 cp -c xamarin-macios/tests/sampletester/bin/Debug/tmp-test-dir/execution-logs/*.xml "$DIR/"
 cd "$DIR"
 git add .
